filling out bank holiday service class

+++ src/main/java/com/leon/rfq/services/BankHolidayService.java
@@ -7,10 +7,10 @@ import com.leon.rfq.common.EnumTypes.LocationEnum;
 
 public interface BankHolidayService
 {
-    int calculateBusinessDaysToExpiry(LocalDate startDate, LocalDate endDate, LocationEnum location);
-    int calculateAllDaysToExpiry(LocalDate startDate, LocalDate endDate);
-    int calculateAllDaysToExpiryFromToday(LocalDate endDate);
-    int calculateBusinessDaysToExpiryFromToday(LocalDate endDate, LocationEnum location);
+    long calculateBusinessDaysToExpiry(LocalDate startDate, LocalDate endDate, LocationEnum location);
+    long calculateAllDaysToExpiry(LocalDate startDate, LocalDate endDate);
+    long calculateAllDaysToExpiryFromToday(LocalDate endDate);
+    long calculateBusinessDaysToExpiryFromToday(LocalDate endDate, LocationEnum location);
     boolean isBankHoliday(LocalDate dateToCheck, LocationEnum location);
     boolean isValidBusinessDay(LocalDate dateToValidate, LocationEnum location);
     boolean insert(LocationEnum location, LocalDate dateToBeInserted, String savedByUser);
